+++ packages/toolkit/config/webpack/target.js
@@ -1,3 +1,11 @@
-module.exports = ({ defaultTargets }) => {
+module.exports = ({ defaultTargets, packageConfig: { target, packageType } }) => {
+	if (target) {
+		return target;
+	}
+
+	if (packageType === 'module') {
+		return 'es2020';
+	}
+
 	return `browserslist:${defaultTargets.join(', ')}`;
 };