+++ projects/library-ts/src/util/index.ts
@@ -1,0 +1,3 @@
+export function sayUtil(name: string) {
+	console.log(`Hi Util! ${name}`);
+}